# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rota Certa Passagens - API do site público
  version: 0.1.0
  description: Contrato da API do site público e do painel administrativo.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rc_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    cronToken:
      type: http
      scheme: bearer
      description: NOTIFICATIONS_CRON_TOKEN, for unattended scheduler calls with no session cookie.
...
paths:
...
  /api/admin/finance/mileage-lots:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista lotes de milhas (Fase 5), com balanceQuantity e isExpired calculados no servidor
      parameters:
        - { name: counterpartyId, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: program, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [active, depleted, expired, canceled] } }
        - { name: currency, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Lista de lotes }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Compra um lote de milhas; cria automaticamente, na mesma transação, a obrigação de pagamento ao fornecedor (custo direto)
      responses:
        "201": { description: "{ id, obligationId }" }
        "422": { description: category_must_be_direct_cost ou vínculo inválido }
  /api/admin/finance/mileage-lots/{id}:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Detalhe do lote, incluindo todas as alocações (ativas e estornadas)
      responses:
        "200": { description: "{ mileageLot, allocations }" }
        "404": { description: Não encontrado }
  /api/admin/finance/mileage-lots/{id}/cancel:
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cancela um lote sem nenhuma alocação ativa — nunca DELETE; lote com alocação nunca pode ser apagado nem escondido
      responses:
        "200": { description: "{ ok: true }" }
        "409": { description: mileage_lot_has_active_allocations }
  /api/admin/finance/issuances/{issuanceId}/mileage-allocations:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista as alocações de milhas de uma emissão (pode ter mais de um lote, emissão híbrida)
      responses:
        "200": { description: Lista de alocações }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Aloca milhas de um lote para a emissão; nunca permite saldo negativo mesmo sob concorrência (SELECT ... FOR UPDATE no lote); recalcula miles_quantity/miles_cost_cents da emissão
      description: >
        Só permitido enquanto a emissão está `pending` (mesma trava de custo histórico da Fase 4).
        O custo da alocação é calculado com BigInt a partir de total_cost_cents/quantity_purchased
        do lote (shared/mileageCost.ts) — nunca float, nunca reprecificado depois. Prova de
        concorrência real (duas transações Postgres genuínas disputando a última unidade) em
        tests/mileage-allocation-concurrency.pg-real.test.ts, gated e não executada por `pnpm test`.
      responses:
        "201": { description: "{ id }" }
        "409": { description: insufficient_mileage_balance, mileage_lot_not_active, mileage_lot_expired ou issuance_locked_after_issued }
        "422": { description: issuance_mode_does_not_use_miles ou mileage_lot_currency_must_match_issuance }
  /api/admin/finance/mileage-allocations/{id}/void:
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Estorna uma alocação (devolve o saldo ao lote, reabre o lote se estava esgotado) — nunca apaga a alocação original; no máximo um estorno por alocação
      responses:
        "200": { description: "{ ok: true }" }
        "409": { description: allocation_already_voided ou issuance_locked_after_issued }
...